+++ lib/index.ts
@@ -8,10 +8,12 @@ export { useGridCallbackRef } from "./components/grid/useGridCallbackRef";
 export { useGridRef } from "./components/grid/useGridRef";
 export { List } from "./components/list/List";
 export {
+  type DynamicRowHeight,
   type ListImperativeAPI,
   type ListProps,
   type RowComponentProps
 } from "./components/list/types";
+export { useDynamicRowHeight } from "./components/list/useDynamicRowHeight";
 export { useListCallbackRef } from "./components/list/useListCallbackRef";
 export { useListRef } from "./components/list/useListRef";
 export { type Align } from "./types";