+++ src/nav/Nav.tsx
@@ -8,6 +8,7 @@ export function Nav() {
       <NavSection label="Lists">
         <NavLink path="/list/fixed-row-height">Fixed row heights</NavLink>
         <NavLink path="/list/variable-row-height">Variable row heights</NavLink>
+        <NavLink path="/list/dynamic-row-height">Dynamic row heights</NavLink>
         <NavLink path="/list/props">Component props</NavLink>
         <NavLink path="/list/imperative-api">Imperative API</NavLink>
         <NavLink path="/list/aria-roles">ARIA roles</NavLink>
@@ -25,7 +26,7 @@ export function Nav() {
       <NavSection label="Other">
         <NavLink path="/grid/rtl-grids">Right to left content</NavLink>
         <NavLink path="/grid/horizontal-lists">Horizontal lists</NavLink>
-        <NavLink path="/list/images">Variable size images</NavLink>
+        <NavLink path="/list/images">Images</NavLink>
         <NavLink path="/list/sticky-rows">Sticky rows</NavLink>
       </NavSection>
       <div>